requirements.txt
fastapi>=0.110.0
uvicorn==0.24.0
mcp-server>=0.1.4
mcp-python-sdk>=1.6.0
requests>=2.31.0
sqlite3